requirements.txt
ollama>=0.6.1
chainlit>=2.0.0
python-dotenv>=1.0.0